IJPL-1002 Rename TreeState visitors
The name Visitor was too generic, and the name BulkExpandVisitor
was too specific. The real difference between the two is that one
visits one path only and the other one visits multiple paths.

We're going to use MultiplePathsVisitor to load cached nodes
now, so the old name doesn't make much sense anymore,
as they're already expanded by that time, just not loaded yet.

GitOrigin-RevId: 397710fbf758f83bc7fc8a7adc45d2de689f94c7