[platform] use default implementation of MarketplaceCustomizationService if application isn't loaded yet (IJPL-149450)
During importing settings from a previous version, ConfigImportHelper may start downloading of new versions of the plugins installed in the old version of the IDE. At that moment Application instance isn't created yet, so services aren't available, and attempt to obtain a service results in NullPointerException. As a simple workaround, the default implementation is returned in such cases.

GitOrigin-RevId: 4148654662bfb88e1bdc7e6a1cd15d259a2882e2