[RDCT] RDCT-1224: Remove @RemoteSetting annotation and replace it with a static map in RdSettingsStorageService
There are only a handful of PersistentStateComponent that we want to synchronize. After a discussion with @develar, we decided not to have an extension point and special annotations in the platform, but to list all of such components in our code.
Components' names will not be changed to keep backward compatibility for XML settings, so it is a safe approach. In the future, we will sunset PSC anyway and migrate them to the new SettingsController API.

IJ-MR-130819

GitOrigin-RevId: 7c9b974d406f8f4179d671f6483fd7143b6618d9