IJPL-43545 Use native "normal" bounds on windows to save frame state
When the IDE frame is maximized at the moment we close it,
we have to resort to all sorts of tricks to figure out what the frame's
"normal" bounds should be, because Swing doesn't report that.
These tricks are not very reliable, especially in multi-monitor configurations,
and especially on Windows where it's easily possible to move a maximized window
with shortcuts.

Fix existing issues and potential ones by calling the native API
instead. For now do it with reflection and JNA, but if it really helps,
we can migrate later to a proper JBR API (which we'll have to create first).

GitOrigin-RevId: 59a3f36a246654275137c40bffc3cff9dd4b6188