[terminal] IJPL-101589 Create code-based command spec for cd command
There is a problem with interpretation of json-based command spec for `cd`. It specifies that there should be directory suggestions + some additional hardcoded values (`-` and `~`).
So, our completion engine is suggesting hardcoded values even if user already typed some path. For example, we should not suggest `-` and `~` in this case: `cd src/<caret>`.
There is no easy way of solving it on the completion engine side, so better to fix in on the client side - create our own spec for `cd`, that takes into account this problem.

GitOrigin-RevId: c4231769932d66e582a5234eb12d0c25f79a9c80