[Workspace Model] Remove double loop while modules initialization
Each module bridge was creating a listener that processed all new modules.
This is not necessary as the bridge processes only its own module.
To remove the double loop, a separate service was introduced.

The performance issue on this double-loop was not really visible,
but in the case of bazel project that can easily have 25.000 modules
this starts to be noticeable.

This is a part of BAZEL-317

GitOrigin-RevId: 000160abaaed0dca99b86f8f023fbf6fe490f402