IJPL-43545 Introduce minimum size for restoring the frame size
When the IDE frame size is saved, it's possible that the saved values
are incorrect due to some bug or because the storage was corrupted
externally somehow.

When we restore these values, if they're too small, it creates a serious
usability problem, as the resulting frame is very hard to even resize
back to some sensible size.

To fix this, introduce the minimum values. The exact values don't matter
as long as they're smaller than anything sensible that the user actually
may want to save, and as long as they're large enough so that the resulting
window may be easily moved and resized manually. 350x150 seems to work fine.

GitOrigin-RevId: 30edaeba3eb64c05a5e105fc01b33121d5dfc55e