[rdct] use 'frontend' subdirectory instead of generic 'tmp' subdirectory for temp folders in the frontend process (GTW-2077)
If the frontend process uses the same 'idea.paths.selector' as the monolithic IDE, it's better to put its temporary config and system directories under a special 'frontend' folder. This way they won't be mixed with other per-process folders, and we can be sure that they won't be cleared by the monolithic IDE while the frontend process is running.

GitOrigin-RevId: 5662f7fce327bd5c4d38a829d21d420caee63777