IJPL-1002 Create coroutine-based Invoker implementation
Note that it works a bit differently from the legacy one.

One obvious thing is that threads can now be reused,
so it's no longer guaranteed that two background invokers
will always use different threads.

One less obvious thing is that we now have larger read actions,
as the entire invokeSafely is now wrapped into a read action
that is automatically restarted. It shouldn't have any negative
consequences because the code still checks for obsolete and canceled
commands and will not attempt the actual execution unless needed.
So in the worst case, we may end up with a read action that does nothing.

The dispatched/undispatched thing along with yield-or-no-yield
stuff seems to be caused by some quirks in the coroutine implementation.
If we always follow the start-undispatched approach, then yield()
sometimes takes too long (over 100 ms). So only do all this if we actually
need to simulate a single-threaded invoker with strict task ordering.
For multithreaded pools, there are no ordering guarantees to begin with,
so it's safe to just launch and run.

For the EDT, don't use a Semaphore at all, because it screws up
modality. Modal calls wait forever until some code up the stack
finishes, and that code waits for a modal call to finish,
e.g. in a unit test or under modal progress.

GitOrigin-RevId: 20a202694e2770159ab398a226cba172fc6b96f6