[new project wizard][IDEA-340801] fix deadlock on "Create new project from existing sources"
`ProjectFromSourcesBuilderHelper#commit` is executed on the EDT. Deeply inside the `commit` method, a call to `VirtualFileSystem#refreshAndFindFileByPath` is made. This leads to a subsequent call to `RefreshQueueImpl#queueSession`. As a result, we are trying to run a `ReadAction` from a context that does not provide any information about the current modality. And because `commit` was invoked from the EDT, there is no way to execute that `ReadAction`.

GitOrigin-RevId: d9ffe03bd71ae05f0dd6966fe6f2f3b5a315a9c7