[RDCT] RDCT-1224: Fix EditorSettingsExternalizable not synchronizing default settings
Every PersistentStateComponent that overrides some meaningful `loadState(state)` method must override `noStateLoaded()` method as well, because it is called when no settings are changes, resulting in an empty settings state. If `noStateLoaded()` is not overloaded properly, then returning of all values of some settings component to the default values will not be synchronized.

IJ-MR-130819

GitOrigin-RevId: ce8f8290484135b195852d5bdcd0890e1c8c28bc