[platform] reverting 51d24869 (fixes IJPL-34819)
Reverting the commit "[platform] correct quoting of collected VM options in Windows .bat launcher (IDEA-291413)", because the corrected quoting was rather incorrect: it worked for quoted strings, but failed with unquoted ones. Since unquoted strings are the norm (.exe launcher doesn't require quoting, and users should not think about it), the original variant is restored.

GitOrigin-RevId: 6f9e4039abea8ec410ed4f3d1e48c7c7891c1dd4