extract `intellij.platform.clouds`
This is a pure v2 module.

It's needed to solve the following problem:
- `intellij.platform.remoteServers.impl.xml` is xincluded in a bunch of product plugin XMLs
  => it's a part of core class loader;
  => it cannot reference other non-core v2 modules.

But its actions need to reference `intellij.platform.execution.dashboard` (to be extracted).

The plan is to keep `intellij.platform.remoteServers.impl.xml` and xincludes,
but remove everything that is not API from it.

This module now contains actions (thankfully unused from third party plugins),
and it depends on original `intellij.platform.remoteServers.impl` module.

Additionally, `DefaultRemoteServersServiceViewContributor` is moved there because
`intellij.platform.execution.serviceView` is also about to be extracted.

GitOrigin-RevId: e49fea548ce0134e2960c6f64fd121d717e83f0c