[RDCT] RDCT-1224: Fix per-client persistent state components
Initially, persistent state components were not registered under remote clientId. It is not correct, though, because a non-per-client component may be loaded for the first time under a remote clientId, and it must be registered properly. This commit adds an explicit marker for per-client components to correctly implement this behavior.

IJ-MR-130819

GitOrigin-RevId: d170176f400902616e1448b0ecdf73a4c4918e8d