RDCT-1149 '+' are disapearing on hover
Make sure focus events are dispatched, and painting is done with controller's client id in RD setup.
Focus events are explicitly dispatched with controller's client id, as it doesn't seem possible to propagate client id via 'requestFocus' requests and such without introducing a dedicated JBR API.
As for the painting events, we just make them covered with the existing logic of client id propagation for InvocationEvent-s.

GitOrigin-RevId: 52d59e0ffef69dfa4dff7618a8e50551b2a26549