IJPL-797 'intellij.platform.usageView.impl' review internal API
Fix exposures.

We have a strange situation here: UsageNode and GroupNode
are public classes, and yet they extend the package-private
Node class. Not much we can do about this as long as these
classes are used by plugins, which they are.

Most exposures are easily fixed, but there's the UsageNode
constructor that's actually used in a couple of plugins.
They pass null as the parent, though, so create an overload
for them and make the full constructor @Internal.

As for UsageGroupingRuleProviderImpl.createGroupByFileStructureAction,
its signature had to be changed to not expose an internal class.
This will break binary compatibility, but not source compatibility,
because the plugins using this function just add the result
to a DefaultActionGroup, which is a safe thing to do.

GitOrigin-RevId: 12197ea309d3d45ec9e9d5947e533345ba48b439