IJPL-1002 Refactor coroutine Invoker implementation to improve readability
According to IJ-CR-134238, instead of splitting the logic of offer(),
make different implementations of it, duplicating bits of code
if necessary. Now we have three different implementations
for the EDT, "single-threaded" BGT and "thread pool" BGT invokers.

Rename *Impl to *Delegate, use Boolean instead of ThreeState
for useReadAction wherever possible.

GitOrigin-RevId: e505f6749344139bd52a18554209cc15c9334416