IJPL-1037 Fix the Default layout in the layout quick switcher
When updating the actions to non-toggleable to avoid
multi-select issues, the restore factory default action
was left unchanged. So it's still bugged and uses the wrong
"current" indicator (a cherkmark instead of an arrow).

Fix by using the already updated action class,
but delegate the text and description to the "regular"
toggleable action.

GitOrigin-RevId: 3b5812fa8a51b3d88f4b9c4de1609b15e836c188