IJPL-1037 Turn off multi-choice for the Default tool window layout
This isn't a multi-choice menu, once the layout is selected,
it's applied and the menu should be closed. Other layouts
are selected using sub-menus, but the Default action
keeps the parent menu open.

Fix by explicitly setting isMultiChoice = false for
the template presentation.

Note that this would also fix the issue with selecting
the Default layout in the Quick Switcher, but
it would still have the wrong icon (a checkmark
instead of an arrow), so the Quick Switcher
still needs a separate action for the Default layout.

GitOrigin-RevId: a2805f114d6630d10ac6fe177190fe8c12b18568