[Workspace Model] Use multimap instead of bidirectional multimap in library tables listener
Library tables listener keeps an index of module <-> library dependencies.
However, there is no need to use the bidirectional map as the libraries of module can
be easily obtained from the module itself. Also, this direction of index (get libs by module)
was not actually used in code.
Populating bidirectional index is slower than population of a regular map, so we can
replace the bidirectional map with a regular one.

This work is done in the scope of BAZEL-317

GitOrigin-RevId: 3465673c2aa9fef9c4d2ad0e54691750d4a95cba