[java-mrjar] Treat classes with the different multi-release version as equivalent
It looks like this is a better alternative for now. In particular because we don't know how to resolve a reference from a library class within the same library. The result depends on in context of which module we are doing this resolve, but in general we don't know from which module we started. Currently, we resolve to the same version as the original class, but this might be incorrect. If we consider different versions of a class as different classes, we have problems with code-insight features, as for example, return type of a method and its direct supermethod might be non-equivalent, if the method is declared in source file but a supermethod is declared in the library.
Fixes IDEA-353000 must either be declared abstract or implement abstract method

(cherry picked from commit 3fa59bc6c8510f8dc0bcab63ccf491b4ff90610a)

IJ-CR-134506

GitOrigin-RevId: 2abe483fd3a0eee129e98c0a6cfbd3d44c83edfb