[Workspace Model] Avoid unnecessary library resolve in ModuleDependencyIndex
There is no need to resolve the library on every update of the project model as we need to add listeners on the library only of "first appearance" of the library in the project.
If we resolve the library bridge on every update of the project model, this starts to be visible on the performance snapshots. This small refactoring is supposed to improve the performance.

This change is done in the scope of BAZEL-317

GitOrigin-RevId: fac5802dd23168c7ab3ae6478eebe584ba68c383