[github] Add parsing and interpreting of PR/issue IDs (IJPL-149900)
#IJPL-149900 Fixed

Similar to the ones added for GitLab, this adds functionality
to parse pull request links and mark them.

Sadly issues and pull requests share the same ID indicator
for Github, so we can't be fully sure from the ID only whether
we're dealing with a reference to one or the other.

Likely issue: if an issue ID parser is configured, it will likely take precedence
over the pull request ID parser, so it will just try
to open some issue in web, probably.

GitOrigin-RevId: b0193d4c36f08b2c6c1e0096829b843fd0a02b58