[Workspace Model] Remove `LibraryImpl` implementation of the library
Now libraries should work via the Workspace Model with `LibraryEntity` and `LibraryBridgeImpl` as a compatibility API

There are no usages of LibraryImpl (and other removed stuff) in the platform or in the external plugins.

The reason that this implementation is removed now is that LibraryTableBase contains a call to `afterLibraryRenamed` function. This function, in turn, seems obsolete as it updates the references in module dependencies. Such references, however, should be automatically updated by the workspace model feature of automatic update of SymbolicIds.

This change is done during the work on BAZEL-317.

GitOrigin-RevId: 9c89a3c6aefac8c8cf8782074081d763498b0e41