[terminal] IJPL-101589 Provide resolved descriptions to shell completion suggestions
Suggestions are created in `ShellRuntimeDataGenerator`'s, so they are already lazy. There is no need to make the descriptions lazy because the lifecycle of `ShellCompletionSuggestion` object is short.
Also got rid of `ShellCompletionSuggestionBase` as not all implementations of it need to have a lazy description. It would be clearer to just move it to `ShellCommandSpecImpl` and `ShellOptionSpecImpl`

GitOrigin-RevId: ac9957751bcf0cc95bfba01b7af7ba9b73d35b61