IJPL-149966 Fix submenu usability helper
The existing implementation worked the following way.
When a submenu is opened, the mouse position is remembered,
and then all mouse events outside the opened submenu are ignored
as long as the mouse cursor is within the triangle formed
by the initial mouse position and the closest vertical edge
of the submenu. Such a location of the mouse cursor was interpreted
as the mouse moving towards the opened submenu,
and therefore opening other submenus was undesirable,
as that would close the opened submenu before the mouse reaches it.

This approach had a usability issue: if the opened submenu is high
enough, and the mouse cursor was very close to the left edge of
the parent menu, then it was very hard to select the next submenu
intentionally, as almost the entire surface of the next submenu element
was covered by this triangle.

Fix by introducing a mechanism to detect the mouse cursor actively
moving away from the opened submenu. Remember the horizontal
position closest to the opened submenu so far, and when the current
position is further than that by a certain threshold (currently hardcoded
as 16 user-scaled pixels), then it's interpreted now as the user
actively trying to move the mouse away from the opened submenu.

Now if the mouse moves more or less steadily towards the opened
submenu, then the behavior is the same as before. But it's now enough
to move it left/right a bit to disengage this usability helper.

GitOrigin-RevId: 33d1f76f32ca5a813fe88a28855dfc3e0f065fd3