[Workspace Model] [IJPL-150046] Use `referrers` function for updating the links in modules
Strictly speaking, the links in modules will be automatically updated by the project model, but this will not work so in case of global libraries. During the global libraries sync, it may happen that remove+add events will be produced instead of "replace" event. So, we have to keep this logic at least for the global libraries

GitOrigin-RevId: 10796c11aaf9ed296cc5ac2ec90193412c4dba82