IJPL-132292 Only temporarily set search scope if it's not set by the user
For some searches, such as searching for usages of a private method,
global search scopes make little sense. For such searches, the scope
is reset to an "all" scope (Project Files and Libraries). But this scope
is saved in the options, and the next time the settings dialog
is displayed for a global search, the selected scope is this default one.

Fix by storing the previously selected scope temporarily while
the scope is reset. The next time the dialog is created,
reset it back so that the proper search can be displayed.

GitOrigin-RevId: 389bd9a4958fad042ca1d6363e1b1ad2d819dacb