IJPL-132292 Use the previously saved scope in the Find Usages dialog
This effectively makes the dialog ignore the scope setting
provided in the settings. However, this is in line with
the requirement that the previous scope should always be reused.
Other settings are still context-dependent (field search settings,
method search settings, class search settings, etc.) because
there are different settings for different contexts. But the scope
is one common thing between them and therefore it makes little
sense to have individual scope settings for different contexts,
at least as long as we agree that the scope should be saved and reused.

The scope is saved when the dialog is initialized and the selected
scope is applied. It is also still saved in the context-dependent
settings as those can be later reused by Show Usages.

GitOrigin-RevId: 853895d5a3ac47370a2045226747c5db19aa1b8d