IJPL-149923 Fix the Abbreviate Qualified Names option appearing
A regression caused by making it always visible for applicable panes.
A pane was considered applicable if
pane.supportsAbbreviatePackageNames() == true
but apparently it's not enough, must also check
flattenPackages.isEnabled(pane) == true
(as opposed to flattenPackages.isSelected()).
This is because these two are closely related,
and if one is not visible, the other should be hidden as well.

We still show this option regardless of whether
the Flatten Packages option is selected, just disable it
if it isn't applicable in the current state (as opposed to
the current pane/product).

GitOrigin-RevId: 60aecf19496a9c3876abae524934e63b3c8654ce