[lvcs] support different modes when computing differences between directories
When a single revision in the directory history is selected, differences are computed with the local directory version. This will return all the changes made in the revision, plus all the changes made afterward. This commit adds support for comparison with the next revision, to see only changes made in the revision.

IJPL-85116

GitOrigin-RevId: 8c09bfdda925c859cbb226467ebfe4b63854eaf1