IJPL-60317 Only scroll configurables once
The scrolling was just introduced recently. This is an attempt
to make it less annoying and still make it work for the most
important case: when the settings dialog is opened with
a predefined search string.

Instead of constantly scrolling on every search change,
only scroll a given configurable once since it's selected.

GitOrigin-RevId: d99baa608690cc76cca9ffe0d56503069eb07897