[Kotlin] Support collecting of view names in case of non-const val Kotlin property
...references from the side of UAST, because support for evaluation
of such expressions from the Kotlin Compiler Analysis API will be dropped in the next
commit. So now it's an API use-site responsibility to implement an extension to const
evaluation, if they want to.

This logic is needed for Spring MVC Plugin (see tests from previous commit).
It's not clear whether it is needed for other UAST users, but removing it would
be a breaking change, so it's implemented uniformly instead of implementing
in Spring Plugin only.

^KT-68327

GitOrigin-RevId: c4cf9fca3d0ea5d39454b7dfda2d1375edadaf41