IJPL-132292 Fix the default search scope for Find Usages
The default scope names in FindUsagesOptions and in FindSettingsImpl
are different. FindUsagesOptions sets it to "Project Files" while
FindSettingsImpl uses a scope that doesn't exist ("All Project Classes").
This in turn causes the scope chooser combo to simply select
the first item, which is All Places, and hence the default becomes wrong.

Fix by checking whether the saved scope is the default, and if it is,
fall back explicitly to whatever FindUsagesOptions considers
the default ("Project Files") as it seems to be the source of truth here.

GitOrigin-RevId: caccb6423f5d284768eebffe0bde16ab3dbaa6f2