[platform] update `LibraryTableTracker` to work with `LibraryId`s for equality instead of `Library`s
The previous implementation of `LibraryImpl` had an `equals`/`hashCode` implementation, but `LibraryBridgeImpl` does not. This means that the leak tracker incorrectly warns about libraries that have been removed and re-added. Using `LibraryId`s for equality is not the exact same as the previous `equals`/`hashCode` implementation (as it only considers the name and table), but there is an expected uniqueness from a library's id, so it is a reasonable replacement.

GitOrigin-RevId: 2e33e5879c1e2ab9e4840c644be000d20dfcffa8