RUBY-32217 RUBY-32923 Fixed modality issue for sdk refreshers
From now on refresh process has its own progress indicator it runs under.

Any refresh request cancels running refresh if requested  with dominating modality. Otherwise it just waits and starts after current one is finished.

Ruby sdk configurable now re-requests sdk refreshes on configurables opening, and this is expected to restart the refresh process (this does not work as intended for now, presumably because of some platform issues)

IMPORTANT NOTICE: if dialog has its own model for the jdk table which is commited when dialog is closed, it will overwrite all refreshers discoveries and we are not handling this in any way. The real known case is project structure in Intellij IDEA.

GitOrigin-RevId: 66e5305ad94269f48272f65f1bffaa3946472c1c