IJPL-149442 Fix Run Widget action button border update
When Compact Mode is toggled, action button border
insets must change accordingly. We already have
auto-updating insets, but in this case, the logic
for calculating them is special, as it takes the toolbar
orientation into account.

Fix by using the existing JBInsets API accepting a Supplier,
and making it @Internal for this purpose. Replace integer
values with the appropriate suppliers, and JBEmptyBorder
and JBInsets will take care of the rest, updating and scaling
properly whenever needed.

GitOrigin-RevId: a5258b0809c9a571ba93b870f470b3cfaf446645