[RemDev] [IJPL-150367] Bring back the compatibility with IdeaVim plugin by returning `editors(): Stream<Editor>` API
Despite the fact that this API is internal, it's used in IdeaVim plugin. The break of this API caused VIM-3439

The suggested approach is to bring the API back and investigate if we can remove this API from the plugin. As this API is internal, after IdeaVim will not depend on this API, this change can be reverted to have nicer function names.

GitOrigin-RevId: 627193bb20e0b6931bcedf8c86dcba237d0fc9f0