IJPL-150383 Fix getCacheableFile() slow op for ResourceBundleNode
Implement getDefaultVirtualFile() in ResourceBundleImpl
to retrieve the virtual file without resolving the PSI pointer.

Unfortunately, can't do the same for CustomResourceBundle,
as it's not based on smart pointers, so we have to return
null for those. Such bundles correspond to several
files anyway, so there seems to be no reliable way to just
cache a single file name.

GitOrigin-RevId: 969a424df5cf55289e06016502c94569e05e0232