[RDCT] GTW-8489: Add method to `TerminalExecutionConsole` which flushes all text events to the listeners immediately
Text changes are fired to the console listeners on the other thread (EDT), and with a 200ms delay. So there is a possible situation when a console listener gets an event about the text change, that is already applied in the `screenBuffer`. To avoid this behavior, a new `flushImmediately` method was added. To keep the contract, it must be called only from EDT.

GitOrigin-RevId: b1a40ae587fc795a1fc80a03b0e81b8d1b604056