[terminal] IJPL-155919 Add prefixReplacementIndex property to ShellCompletionSuggestion
It is needed to allow clients to specify that their suggestion is for a shorter prefix than the initial.
It will allow removing the heuristics of assuming that typed prefix with a path separator is a file path.
Also, it will allow having subcommands and option names with path separators, that is crucial for `php`.

GitOrigin-RevId: b50afe02d109dff16e9bbd406f6678244322552b