IJPL-48479 Fix slow ops in SelectInTargetPsiWrapper
Not a perfect solution, as it's yet another NBRA seemingly for no reason,
but it's the best we can do without changing the API and enforcing
strong EDT/BGT contracts on it.

Existing getSelectorInFile() implementations seem to be safe to
call from a BGT, there's mainly PSI stuff there and also some access
to thread-safe editor models like the caret position.

GitOrigin-RevId: 92bb27afda5989e283bc98d6a577e11666b21579