IJPL-60987 Fix slow op in ShowUsagesTable
The slow op was caused by copying UsageInfo, which meant
resolving PSI pointers and creating new ones. It doesn't seem
necessary in the current implementation, as UsageInfo objects
are immutable for the most part, and all mutations that do happen
do so before the object is published. Moreover, the copy()
implementation was incorrect: the class has over 400 subclasses,
and none of them overrides copy(), which means the actual type
is lost. It's probably not relevant at this point, but still, it seems wrong.

GitOrigin-RevId: a678bcce74eea1f8bc0ac3dd348a1a92263a2383