IJPL-797 `intellij.platform.ide.util.io` review internal API: mark `LocalPtyOptions` + `Builder` as public API
Getters and setters of `useWinConPty` are marked as internal. They are not supposed to be called by third party plugins.

There was found only one usage of `useWinConPty` in a third party plugin, and it looked like:

```kotlin
builder.useWinConPty(shouldUseWinConPty()).build()
```

Now `shouldUseWinConPty()` is the default value of `useWinConPty`, no need to set it explicitly.

GitOrigin-RevId: 3e64ed7c56ad4ffa68b23b6723122b11cdb09182