[git-stash] show the stashed version as the after version in the stashed diff
The stashed version should be considered as the after version in the diff, as it contains the modifications to be applied to the local version. This behavior is consistent with the diff for the shelves and with the previous implementation. Additionally, showing the local version as the before version would allow displaying a correct background for added or deleted files.

IJPL-155908

GitOrigin-RevId: 59bb2d41b7c30c9d43acdf82ff37d2f6074647fa