IJPL-58764 Do not process input events in popups if there's a window above
A rare, but important case when a modal window opened on top of a popup
keeps it open, and yet the popup receives some input events that should
belong to that window, and that window alone. If that's a password input window,
we have the password leaking to the popup into its speed search or wherever else.

Fix by ignoring all input events in popups if there's a modal window on top of it.

GitOrigin-RevId: ce770186859947b35afb7c3f9b55427dd225f2a8