IJPL-48490 Fix slow ops in ScopeEditorPanel
Migrate to AsyncTreeModel, just like with the Package Dependencies panel.

One difference is that this model needs to sort children, so introduce
a boolean parameter for that to distinguish between this and package dependencies.

Another important note is that instead of a regular NBRA, we use the model's
Invoker to retrieve package sets from the selected nodes. This is because the tree
model is single-threaded by its nature. The read lock would prevent
concurrent access, but it's still hard to reason about.

GitOrigin-RevId: 2095337ecceed015980212341e0075243d9c257e