KTIJ-28554 [kotlin][parcelize] Enable default implementation of `ParcelizeAvailabilityProvider`
It works in JPS, and enables `IdeParcelizeResolveExtension`
when there is a corresponding compiler plugin jar (namely "parcelize-compiler.jar")
specified with the `-Xplugin=` compiler argument

Currently this feature is under registry flag and disabled by default;
it should be enabled by default when it becomes clear that it
works as intended and does not break anything, and then the registry
flag should be completely removed.

Side note: `getInstance` function from `KotlinFacet.kt` returns
`compilerArguments`, but they do not include the arguments from the
"additionalArguments" sections in the Kotlin Facets.

This most probably means that currently the use-sites of this function
are not functioning properly, and need to be checked later - for example,
`KotlinSerializationEnabledChecker`

GitOrigin-RevId: 7fc0ff4dc07cf12770e004ccde93314a6b365ee2