[Project Structure] [IJPL-10393] Speed up the `ModuleEditor.getModule` by removing the unnecessary loop
This loop was joining with the other loop and caused performance issues on project structure dialog open in case there are a lot of modules. See IJPL-10393

However, this loop is NOT needed as it was added for a specific case that doesn't exist anymore. See the comment with the details about why this loop was introduced and why it's not needed anymore.

https://youtrack.jetbrains.com/issue/IJPL-10393/Opening-Project-Structure-in-20k-module-project-causes-freezing#focus=Comments-27-9879534.0-0

GitOrigin-RevId: f8eb0049068f9cc7d0a0d98acd318f2b710df40a