ML-based ranking in the plugin manager
(MP-6452) Fix failing project structure and packaging tests

(MP-6452) Make sure that machine learning score is calculated correctly for each plugin

Previously, there was a cache with mutable list of features. There were two issues with it. First, because of mutability, it could store mlScore for a different query sometimes. Second, there were no guarantees on consequent call of ranking and features logging before the next ranking event happens, so the feature cache could be inconsistent in certain scenarios.

(MP-6452) Introduce search index in plugin manager events that corresponds to actual query order

I also change the definition of plugin manager session. The session is restarted only in two cases: when a user clicks on the Plugins section in settings or when the search is restarted after a plugin installation.

(MP-6452) Add additional features: is result ordered by ML, is user internal, experiment group and version

(MP-6452) Fix missing date of latest plugin update, add days since latest updates and textual features

(MP-6452) Apply suggestions from the code review

(MP-6452) Integrate the ranking plugin with the search and the logging group

In feature extractors, during the first run, I calculate the features for a model to predict plugin relevance and on the second run the features to report to the MP collector (with predicted score, for example). I also implement features cache to calculate the most of the features only once

(MP-6452) Implement the baseline plugin manager session id definition and log it to FUS and MP recorders

I attach the start of a session to enableSearch method of PluginManagerConfigurable that gets triggered on "Plugins" setting group selection in the menu

(MP-6452) Add plugin for plugins ranking in the Plugin Manager

ML in SE: add embedding search integration test subsystem tag


Merge-request: IJ-MR-130364
Merged-by: Evgeny Abramov <Evgeny.Abramov@jetbrains.com>

GitOrigin-RevId: 9136d316aec2ede74bec07798dd8db16e7849f54