[yaml] IJPL-150017 Support commas for missing property fix in YAML hash mappings
The IncorrectOperationException was just a red herring. It's a symptom caused by
the invalidity of the YAML produced by the quick fix when automatically adding
multiple properties.

Basically, when adding several properties in a hash mapping (or "flow mapping"
in the spec terms - the ones with braces that look like JSON objects), the
separator between the properties was an EOL, but it's supposed to be an actual
comma. This led to a broken syntax after the first property insertion.

This was fixed by properly supporting commas for hash mappings.

GitOrigin-RevId: 84056029147a74c397493cc62a1baa58b5616125