Add bytecode inspection for ART in KotlinSyntheticTypeComponentProvider
When trying to detect synthetic methods, KotlinSyntheticTypeComponentProvider uses bytecode inspection to verify that a method simply delegates to a default implementation in an interface. This inspection didn't work for ART, because bytecode on ART is different. This commit implements ART specific bytecode inspection and by that improves stepping behaviour in Android Studio. For example in this test case: `afterDefaultParameterValuesIntf.kt`.