Refactor evaluation-plugin core script
The revision involves an overall code cleanup and improves the script's readability and efficiency. Changes include converting a function constructor to a modern JavaScript class and refactoring the `addDiffView` function to utilize the new `unifiedSlideDiff` function that was created to generate a unified diff view. Unnecessary console logs were also removed for better performance.

GitOrigin-RevId: 870c7c1297474c7c6df5c2e25782a3e93f995f07