Improve code diff view and context handling
The commit adds console logging in the diff function to display changes. It also corrects the handling of CSS classes between different context views for sessionDiv in the evaluation plugin. Specifically, the "diffView" class is correctly added and removed when switching between different contexts. This ensures proper styling and functionality across different components.

GitOrigin-RevId: 6c5cc92c05ab67300d7c8184733d3d1946c00b2b