IJPL-155988 Fire change listeners in ExpandableComboAction on popup show/hide
Since ExpandableComboAction extends the model and overrides
isSelected(), we must fire change listeners when the overridden value
changes as well. Failure to do so would result in an obsolete state
being painted.

The reason why this regression appeared after IJPL-149640 is that
prior to that fix, the toolbar would be updated too often for no reason
(leading to a slowdown), and that would mask this issue, even though
it existed even before the fix. So it's not like fixing IJPL-149640 broke
anything, it has just exposed something that was already broken.

GitOrigin-RevId: ffb364cbc54bb612da2d7a0c0aef8b9c71feed66