[intention-preview] Do not attempt to commit physical files
It's possible that the action was invoked on a file which cannot have a non-physical copy (e.g., binary file). In this case, it's probably unchanged by default preview implementation (e.g., custom-diff or html preview is returned), but in any case we should not try to commit it.
Fixes EA-752756 - REWA: ApplicationImpl.throwThreadAccessException

GitOrigin-RevId: 77f708dea0d7e44d13c103a7b4e2ba465a879ff1