IJPL-156567 Add "nodes changed" swap test for the tree layout cache
There's a bug in DefaultTreeLayoutCache: when several nodes
are updated in a way that makes them exchange node paths
(e.g., sorting), then the path mapping update incorrectly removes
some of the affected mappings. It goes like this: a node removes
its old path mapping, adds a new one, the next node removes
its old path mapping, but it matches the new mapping for the previous
node, so it must not be removed.

This test reproduces this bug.

GitOrigin-RevId: 76b1c13f3b27c10c44995037526cb4156a37ee41