IJPL-156567 Fix children update in DefaultTreeLayoutCache
Because some child may have the new value matching
the old value of some other child, we can't just update
the mappings one-by-one. Instead, we remove all old ones first
and then add the new ones.

Interestingly, this implementation was based on
VariableHeightLayoutCache from Swing, but there
it doesn't break anything, even though the layout cache
ends up in an inconsistent state. This is because it has
a fallback: if it doesn't find a certain path in the map,
it'll look it up by traversing the tree. We don't want to do that,
because it would lead to a performance penalty,
and the whole reason for having our own cache implementation
is to improve performance.

GitOrigin-RevId: 5c2e48459d06fedd1a7e7dffc88f1fcdbef88836