Fix for `Snapshot publication is broken after adding PGP signing #405` (#407)
* Add PGP signing to publications

Introduce Pretty Good Privacy (PGP) signing to the publication workflows of the project. The signing setup is included in the gradle wiring along with the retrieval of the key details from environment variables. This change is applied in both the publish and hotfix workflows.

* Update key retrieval in build script and workflow file

* Add PGP keys to GitHub Actions workflow

The commit adds PGP_PASSWORD and PGP_PRIVATE_KEY to the environment variables in the publication workflow. These additions permit secure, authenticated signing of packages before publication.
GitOrigin-RevId: cdcb05ed8f159d3c880cc306bbb6605f0788f7d2